Updated openstack/openstack Project: openstack/heat  cbc0572038b4e06be543f6f36f39346b58933b96
Ignore string default for software config input

We recently (I87edc0d8f1fba2fb61276c682a60c1b2241b5705) enforced default
type for software config input, whereas it was always a string before.
It causes issues with upgrades, as we used to store "" as a default for
all configs regarless of their type, and it failed when we started
validating.

This identifies this case and skip validation.

Change-Id: I8a4f4e834b0862ecef4b97f208f4b03be3572e86
Closes-Bug: #1697627