Updated openstack/openstack Project: openstack/nova  dbf5fdb156f01196e8994798e8eea7e6b0926697
Explain why API services are filtered out of GET /os-services

This came up in the original review for the change that added
the filtering:

  Icadd380ea1ff75f0cca433b68441ac5dad0ead53

But was never followed up, and it comes up from time to time
as a question about the API behavior. This change puts a comment
in the filtering code to explain it.

Change-Id: Ibb0bb0888e96c17147beb47d0a44f8c65ed0b56c
Related-Bug: #1543625