Updated openstack/openstack Project: openstack/glance  edcca24fba421fd8b71a399b59f87e5b365f4c0d
Add hide hypervisor id on guest host

Image hide_hypervisor_id property helps libvirt set in xml instance file
property, which will hide on guest host KVM hypervisor signature
("KVMKVMKVM\0\0\0").
According to the commit message in QEMU repository [1]:

"The latest Nvidia driver (337.88) specifically checks
for KVM as the hypervisor and reports Code 43 for the
driver  in a Windows guest when found.  Removing or
changing the KVM signature is sufficient for the driver
to load and work."

[1]: http://git.qemu.org/?p=qemu.git;a=commitdiff;h=f522d2a

DocImpact

Implements: blueprint add-kvm-hidden-feature
Change-Id: I825c5c2a8ba3b9f5eccd9800870b6307d201614c