Updated openstack/openstack Project: openstack/governance  312e5107acf7e62aae23f57cbb610d0f45ebaac3
Introduce Top 5 help wanted list

A lot of people, especially contributors and organizations that are new
to OpenStack, are seeking guidance on how to make useful, strategic
contributions to OpenStack. While some of those strategic priorities are
long-standing and well-known, in the absence of a formal list blessed by
the Technical Committee a lot of contributors are turning to
Stackalytics to guess what a useful contribution is. Beyond wasted
resources and being annoying to core reviewers, this results in a
failure to drive those new contributors to what could be extremely
useful contributions, and prevents them to step up in the community.

In order to address that, this change establishes a "top 5" list of
areas where help is wanted. At this stage we'll leave it up to
Foundation staff to promote (through keynotes, web pages and blog
articles) the contributors and organizations who step up to help in
those areas. This list will send a clear signal to folks who are looking
at things like Stackalytics and tell them where exactly their effort is
valuable for the community as a whole. Folks on the Technical Committee
will help guide the work, monitor progress and report back to the wider
audience.

Change-Id: I91a24de4f812319d45709e7c65210c6bc49b3213
Co-Authored-By: Davanum Srinivas <davanum@gmail.com>