Updated openstack/openstack Project: openstack/nova  de8096a59d80d10ff1ccf14e0b345be641ba4f07
Remove code that produces warning in modern Python

When running py27 tests, you will see a bunch of the following
warnings:

BaseException.message has been deprecated as of Python 2.6

Because of some old code in the OpenStackApiException base class that
was setting the message attribute of the base class. This code isn't
necessary.

Change-Id: I24357d3069f653a707f15abb34d9d2546c3ee044