Updated openstack/openstack Project: openstack/tripleo-heat-templates  ffcc946a6bd2a74d003f20fa0f5330073c2ce4ac
Fix rpms being installed via DeployArtifactURLs

The deploy-artifacts.sh script is supposed to support installing rpms
when provided by DeployArtifactUrls. The problem is that it uses yum to
install which does not actually work unless the filename ends with .rpm.
This change updates the script to rename the downloaded file to end with
.rpm if it is an rpm so that it is properly installed.

Change-Id: I048d2b4474f9efe424e98e3868f325704e9c352f
Closes-Bug: #1697102