Updated openstack/openstack Project: openstack-infra/project-config  a956d8a5e2dd5cffa399294c0aaa4eb96c6e8841
ceilometer: do not install Gnocchi via devstack in master

This continues to install Gnocchi via devstack in stable branches, but stop
doing so in master.
The downside is that since we cannot point to the new GitHub repository, we'll
continue to test against old stalled stable/* branches of the old Gnocchi
repository (for now?).

Change-Id: I8e3dbcab268c94a08e0f0e22380c2b5663ecd470