Updated openstack/openstack Project: openstack-infra/project-config  7caad46c566123e48e79506b74eff2a623f2225e
Disable Swift on Kuryr-K8s tempest gate

This commit modifies the upstream gate config in kuryr-k8s so it
disables swift. It specifically sets the enabled services and marks the
KURYR_K8S_API_PORT var so it can be easily identified and tweaked in the
future, as it was colliding with the swift-proxy port.

In addition, it triggers the all-projects for tempest so the kuryr-k8s
plugin tests are loaded.

Change-Id: Ia51a12c36d80c185ebf550cc00ffa4fe469db7c4
Signed-off-by: Daniel Mellado <dmellado@redhat.com>