Updated openstack/openstack Project: openstack/governance  7f59c979591fefe546e74cbeacd9b48d84ea366c
Remove "meritocracy" from the opens

The value and meaning of the word "meritocracy" has been a source of
much debate in many communities, open source and otherwise. For many
people it is considered exclusionary and something of a hallmark for
not being in tune with its negative connotations. Rather than
presenting our defenses against those negative connotations and
being more explicit about the local definition of a subjective term,
we can remove it in favor of something that is more direct.

The intent, in the past, has been that "meritocracy" is supposed to
be indicative of a community where the people who get on and do
things and get stuff done are the ones that become become leaders.
The problem with this is "meritocracy" is defined by those who are
already in positions of leadership and thus implicitly excludes
those who may be leading or attempting to lead (or "do") in styles
that are different from existing leaders or who come from
backgrounds different from the norm. These styles may have just as
much "merit" as any other.

The change here simply states what is the case: one of the ways we
achieve openness is by technical governors being elected from
members of the community, rather than being imposed from elsewhere.

Also in this change "technical leads" is adjusted to "team leads" to
reflect the current expansion of "PTL".

Change-Id: Ifdadaad9e616c0ee3ce59f3c1da323ed458c0bc7