Updated openstack/openstack Project: openstack/tripleo-heat-templates  eda12ae632d0d2756b51868858a87f5452aab85f
Add os-collect-config data as an output

Adds a new output, ServerOsCollectConfigData, which is the
os-collect-config configuration associated with each server resource.
This can be used to [pre]configure the os-collect-config agents on
deployed-server's.

Having the data available as a stack output is more user friendly than
having to query several nested levels of stack resources, and then
inspect resource metadata.

implements blueprint split-stack-default
Change-Id: Iaf062f1a72e2a9e4d97f84c67f72408a6b5cebfc
Depends-On: I8acfd67cd8138d587cc362184c84a08134bf3157