Updated openstack/openstack Project: openstack/tripleo-heat-templates  5dc466725ba47b4d78e2b5869d4df470f262a281
Add DeploymentSwiftDataMap parameter

The DeploymentSwiftDataMap parameter is used to set the
deployment_swift_data property on the Server resoures. The parameter is
a map of role names and node indexes to Swift container and object names
to be used for storing deployment data.

The parameter allows for using predefined Swift objects for storing
deployment data instead of container/object names with generated uuid's
from Heat.

implements blueprint split-stack-default
Depends-On: Ia07e9374a4b95bd0e74fc47fb9df4bf6ad096715

Change-Id: I471037de35e7f349d900462ec3ffb16fe2d6ebd9