Updated openstack/openstack Project: openstack/os-net-config  03d3a5dea1b70c62d585dd6de5e3f9c04345e58b
Continue bringing up interfaces even if one fails

Currently os-net-config exits immediately if an interface fails to
come up.  This causes problems when we call it with a failsafe
configuration because it can result in working interfaces not
being configured if there are also non-functional interfaces, which
can leave a system unreachable over the network even though it may
have a working connection.

This change stores errors and handles them after all interfaces have
been processed to allow os-net-config to do what it can even with an
invalid configuration.  If any failures are detected it will still
cause os-net-config to report failure at the end.

Change-Id: I3bc75e217d0b7c5ae62900f4253ad57ee3720685
Closes-Bug: 1692725