Updated openstack/openstack Project: openstack-infra/tripleo-ci  fcfc01da7ebb38a24eeab1c195311f58c37f41ac
Add a driver-list call to the undercloud sanity check

Running 'ironic node-list' only makes sure that Ironic API is running.
We can detect that a conductor is running by making sure that the driver
list is not empty. As we always enable at least one IPMI-based driver
(it's the default, and required for both OVB and quickstart), we can
check for "ipmi" string in the output.

Change-Id: I2486a1d00f53442eab7dc99ee9fdf716d822c72e