Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  7b4b9dcb8979271fbc78272014bfc24402fcc47b
Add `do-not-fetch-tags` to CloneOption for Git

In one of the recent change to Git plugin behavior[1], tags are
also search as part of branches.

This cause projects that build on all branches to also build on all
tags.

This change allowing for tags to be ignored.

Default this option is false and missing from XML output.

[1] https://github.com/jenkinsci/git-plugin/pull/340
[2] https://github.com/jenkinsci/git-plugin/commit/bfeda3e661531bd9baec937ec4a4944ac482692c

Change-Id: I2ed2290f9ef8ecd0d9e96aad1d7cbce2964bf2da