Updated openstack/openstack Project: openstack/cinder  5d1c7442ea05f5abefd8a8fe3873c5d500833d6f
Remove deprecated 'host' option for cinder-volume

The cinder-volume command has a 'host' option that has been
deprecated for over 3 years. This finally removes it.

Not including release note as this is old code and this command
isn't really something an end user would likey care about. Mentioning
it in a release note may just cause confusion at this point.

Change-Id: I70f6876ff36e63d572e587cd3cb00b70cf434ca4