Updated openstack/openstack Project: openstack/neutron  c6bc9e3dd50cc51acd2d2fe3d130ff6b14407f5d
Fix list QoS rule_types tempest API test

Test was wrong because it was only checking if each of
expected rule types is in list returned from Neutron server.
As list of expected rule types was wrong, it didn't check anything.

This test was also dependent on deployment because it was checking
available rule types returned by neutron-server.

Now it is changed that it's checking only returned keys in each
rule_type and it's not checking returned values.

Change-Id: I467672cc0ba1e579cddfff6020658c82e80fa5af