Updated openstack/openstack Project: openstack-infra/tripleo-ci  58978cb980bb2438042ffe425cf396567e4dd0b4
Set jobs subunit file as fail until it's overwritten

We have testrepository.subunit file in logs for parsing it by
openstack-health. But if job doesn't reach pingtest stage, this
file isn't created and job isn't registered there. So let's
create failing subunit file and then will overwrite it with
tempest of pingtest real subunit in the end of job.

Change-Id: Ic86e2e8873f1de2316a062a7987b563c5d3d093d