Updated openstack/openstack Project: openstack-infra/shade  98a47b8b86264fa3015a5b8c71ee9e07e5c6b78f
Project update to change enabled only when provided

The project update call considers enabled=True as the default, causing
the project to always become enabled unless enabled=False is passed
explicitly.

This patch removes that default and only touches the enable field when
it is explicitly provided as enabled=False or enabled=True.

Closes-Bug: #2001080

Change-Id: I0a3b926b42be0321d06ebc370e4f51eba4150a50