Updated openstack/openstack Project: openstack/glance  0728ff34d2e6c4c1e047d57cecef5e52a17b885b
fix the autodoc instructions

Use autodoc_index_modules instead of autodoc_tree_index_modules so only
one copy of the API docs are generated.

Exclude the tests subpackage because it's not part of the Glance Python
API.

Exclude the glance.db.sqlalchemy subpackage to avoid issues caused when
parts of it cannot be imported during the doc build because they rely on
alembic setting up values that are not configured for a simple import.

Change-Id: I127b812bc879fc39533b39df1785d93835142430
Signed-off-by: Doug Hellmann <doug@doughellmann.com>