Updated openstack/openstack Project: openstack-infra/shade  10e6fbe44f53a0beff2a2e4d2243b3c2e7f96d8c
Make sure we pass propert dicts to validate

In register_uris, validate takes a dict that matches the dict one
normally passes to requests_mock - that is, it can take json, headers,
etc.

Unfortunately, our fixture doesn't error when one passes a normal dict
to it thinking you're passing a json dict for it to validate.

This fixes the call sites - the associated tests need to be fixed.

Change-Id: I16f265fe41219be0f3b031eb8be26f740a65d665