Updated openstack/openstack Project: openstack-infra/shade  d1eea7a720bb14f6781ec5abadbc27fafdf85e1e
Make sure we don't fail open on bad input to validate

Passing arbitrary dicts to validate does not validate that that's what
we're sending to the API. Put in a validation check that the keys we're
sending are the ones validate does something about.

It's possible there is another param to validate we're not listing. If
we hit a test that needs it, we can always add it.

Change-Id: I75b24f4f640b7cf6ffebff494e1627569d74755a