Updated openstack/openstack Project: openstack-infra/infra-specs  647459733e821299b1cba0d17c20a0b15cd5867b
Ansible Puppet Apply is implemented

Per discussion in the most June 6 Infrastructure team meeting:

http://eavesdrop.openstack.org/meetings/infra/2017/infra.2017-06-06-19.03.html

...the Ansible Puppet Apply specification should be considered
completed and so is being moved into the implemented subsection of
the index. This was overlooked in prior cleanup change
Ia603a1f752e0c8acb43ed323864267c2b69bdf03 because it was in the
Priority Efforts section.

Change-Id: Icfe7e3d90349ac6df8f36bcc87c866bea8834bdc