Updated openstack/openstack Project: openstack/nova  1d5e693fff4006f174ff55b0287294a85d13fbaf
Add a reset for traits DB sync

Some functional tests were failing, due to standard traits not being
found. The cause was that there is a flag that ensures that these
standard traits are only loaded once from the os_traits module to the
DB. The test fixtures, though, would clean up the DB between runs, and
the traits would never be re-loaded. This patch fixes that by directly
resetting this flag in the base test code.

Closes-bug: #1701035

Change-Id: I5dc0220310986c5b9fe5b71013c573c3d98cb9aa