Updated openstack/openstack Project: openstack/requirements  251e7a8dee5137f5ca24783473736006a0dad852
Remove mistral from global requirements

Historically, to write Mistral actions you would need to depend on
Mistral itself. However, with the introduction of mistral-lib this is no
longer required. The only consumer of mistral-as-a-library within
OpenStack was tripleo-common. It has been fully migrated to mistral-lib
(see depends-on) and the requirement can thus be removed.

Change-Id: I0a99c19d3cbcfbb396b09c008aff7a379a273d4c
Depends-On: I7ae2e0e4dbca6c02ed8ceedc066ca35eeb13c322