Updated openstack/openstack Project: openstack-infra/project-config  e311282ab940cad68c7e7785d50ac91077faf278
update unified publishing job to work for tags

The old version of the job converted the tag name to a branch and used
that as the publishing location. We do, however, want to be able to
publish docs for a given version of a deliverable when we tag it. That
will allow us to refer users directly to the guide for the version of
their tool or service.

This change assumes that the job is only configured in the release queue
and will not be run for pre-releases. If it sees that the trigger for
the job is a tag, it uses that as the output location. Otherwise it uses
the branch name. That gives us docs.openstack.org/$project/$version or
docs.openstack.org/$project/latest or
docs.openstack.org/$project/$series as the final destination of the
docs.

Change-Id: I5a3fe2775fc21c23fb1df6f6c987ec63000f1c7c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>