Updated openstack/openstack Project: openstack/cinder  429db9d9f37219dfd5a3e1987a49eaa0ba3a5a99
SMBFS: Use share mountpoint when fetching capacity info

One common use case for the Cinder SMB driver is having the service
running within a Windows Scale-Out File Server (SOFS) cluster.

The issue is that when fetching the share capacity info, we're using
the share UNC path directly instead of the local mountpoint. Note that
accessing loopback SOFS shares is not allowed and may fail (depending
on the SOFS address resolution, which may be any of the SOFS cluster
node addresses - if at that specific moment, the address is resolved
to a remote SOFS node address, this operation will succeed, failing
otherwise).

Simply using the local share mountpoint prevents this issue.

Change-Id: I54ccadefdfb33cac749809282a8d9ca00d973cb0
Closes-Bug: #1688225