Updated openstack/openstack Project: openstack/nova  ad6d33997d54d845bbd5571adf9527e56b34e3ed
_schedule_instances() supporting a RequestSpec object

When we merged the use of RequestSpec objects for all the operations, we left
a lot of TODOs for untangling the use of the object with the scheduler.utils
methods instead of the legacy dictionaries.

Some other changes will be provided for changing how we use the objects in the
conductor (and fixing how we pass them to the scheduler.utils) but for the
moment, just modifying the needed setup_instance_group() method to also support
a RequestSpec object.

Change-Id: I434af8e4ad991ac114dd67d66797a562d16bafe2
Partially-Implements: blueprint request-spec-use-by-compute