Updated openstack/openstack Project: openstack/nova  94a2fc8e6330e7618b20067acdb796e27e91f18c
Fix the note at the end of allocate_for_instance

The note changed here was originally added in:

I8f3af96051268503596007491af365e8ce28f5b3

Which introduced a problem which was fixed in:

Ib0d7d91acfe5bd684cdfc072fe47c8115c2f91b3

But the note wasn't updated to reflect the filtering
so it's wrong, and has been wrong for a long time (Havana).

This changes the wording to reflect what it currently
does and in what cases, including an example for the
attach case because the "preexisting" in the variable
name is very confusing since for an attach operation,
it's not *all* of the preexisting ports attached to the
instance, only the ones processed in a given attach
operation.

Change-Id: Iba5082e4e25f5aa748bcbe46ebfc1491b1e575c8